datascience/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
PyMySQL
uwsgi
flask-restful